requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.7
SQLAlchemy-Utils==0.26.16
Werkzeug==0.9.6
ecdsa==0.11
itsdangerous==0.24
paramiko==1.15.1
pycrypto==2.6.1
six==1.8.0
wsgiref==0.1.2
